Cast batch token ids to int64 in the data loader
Token ids are stored in the narrowest integer type that fits the vocab (int16 for vocab < 2**15 via get_unsigned_integer_type), but torch.embedding and the cross-entropy loss require int32/int64 indices. Any dataset with a sub-32768 vocab therefore crashed on the first forward step with "Expected ... Long, Int; but got torch.cuda.ShortTensor". examples/mistral.yaml triggers it directly (vocab 32000).

TokenBatch._get_model_input's meta branch already declares tokens as int64, but the real collation path kept the dataset's narrow dtype. Normalize at the single chokepoint, TokenBatch.from_documents, so both token_ids and the derived labels honor the int64 contract.

+++ fast_llm/data/document/token.py
@@ -77,7 +77,7 @@ class TokenBatch(Batch, TokenDocument):
                 tokens.append(tokens[0].new_full([padding], -100))
                 lengths.append(padding)
         return cls(
-            tokens=torch.cat(tokens),
+            tokens=torch.cat(tokens).to(torch.int64),
             lengths=lengths,
             unpadded_length=unpadded_length,
         )
